Integrated commits from main until b74198e6c3b70c28c0850ab53d1ba48a8e7641cb (I think), and also fixed a bug where the close callback wasn't called all the time, so some files were permanently locked in cache, and also a bug where, if the lowest item in the cache is locked, it wouldn't send anything
