+++ src/header/server.h
@@ -239,10 +239,22 @@ class server<server_type::NON_TLS>: public server_base<server_type::NON_TLS> {
         auto data = new multi_write(std::move(buff), num_clients);
 
         for(auto client_idx_ptr = begin; client_idx_ptr != end; client_idx_ptr++){
-          auto &client = clients[*client_idx_ptr];
+          auto &client = clients[(int)*client_idx_ptr];
           client.send_data.emplace(data);
           if(client.send_data.size() == 1) //only adds a write request in the case that the queue was empty before this
             add_write_req(*client_idx_ptr, event_type::WRITE, &(data->buff[0]), data->buff.size());
+        }
+      }
+    }
+
+    template<typename U>
+    void broadcast_message(U begin, U end, int num_clients, char *buff, size_t length){
+      if(num_clients > 0){
+        for(auto client_idx_ptr = begin; client_idx_ptr != end; client_idx_ptr++){
+          auto &client = clients[(int)*client_idx_ptr];
+          client.send_data.emplace(buff, length);
+          if(client.send_data.size() == 1) //only adds a write request in the case that the queue was empty before this
+            add_write_req(*client_idx_ptr, event_type::WRITE, buff, length);
         }
       }
     }
@@ -292,10 +304,22 @@ class server<server_type::TLS>: public server_base<server_type::TLS> {
         auto data = new multi_write(std::move(buff), num_clients);
 
         for(auto client_idx_ptr = begin; client_idx_ptr != end; client_idx_ptr++){
-          auto &client = clients[*client_idx_ptr];
+          auto &client = clients[(int)*client_idx_ptr];
           client.send_data.emplace(data);
           if(client.send_data.size() == 1) //only adds a write request in the case that the queue was empty before this
             wolfSSL_write(client.ssl, &(data->buff[0]), data->buff.size());
+        }
+      }
+    }
+
+    template<typename U>
+    void broadcast_message(U begin, U end, int num_clients, char *buff, size_t length){
+      if(num_clients > 0){
+        for(auto client_idx_ptr = begin; client_idx_ptr != end; client_idx_ptr++){
+          auto &client = clients[(int)*client_idx_ptr];
+          client.send_data.emplace(buff, length);
+          if(client.send_data.size() == 1) //only adds a write request in the case that the queue was empty before this
+            wolfSSL_write(client.ssl, buff, length);
         }
       }
     }
