+++ src/web_server/callbacks.tcc
@@ -24,21 +24,21 @@ void event_cb(server<T> *tcp_server, void *custom_obj){ //the accept callback
 
   std::string str = "hello world....\n";
   
-  // auto data = basic_web_server->make_ws_frame(str, websocket_non_control_opcodes::binary_frame); //echos back whatever you send
-  // tcp_server->broadcast_message(client_idxs.cbegin(), client_idxs.cend(), client_idxs.size(), std::move(data));
+  auto data = basic_web_server->make_ws_frame(str, websocket_non_control_opcodes::binary_frame); //echos back whatever you send
+  tcp_server->broadcast_message(client_idxs.cbegin(), client_idxs.cend(), client_idxs.size(), std::move(data));
 }
 
 template<server_type T>
 void custom_read_cb(int client_idx, int fd, std::vector<char> &&buff, server<T> *tcp_server, void *custom_obj){
   const auto basic_web_server = (web_server<T>*)custom_obj;
   const auto &filepath = basic_web_server->tcp_clients[client_idx].last_requested_read_filepath;
-  basic_web_server->web_cache.try_insert_item(client_idx, filepath, std::move(buff));
 
-  std::cout << "readcb called\n";
-
-  const auto ret_data = basic_web_server->web_cache.fetch_item(filepath, client_idx);
-
-  tcp_server->write_connection(client_idx, ret_data.buff, ret_data.size); // no need to check if it's found, since we just inserted it
+  if(basic_web_server->web_cache.try_insert_item(client_idx, filepath, std::move(buff))){ // try inserting the item
+    const auto ret_data = basic_web_server->web_cache.fetch_item(filepath, client_idx, basic_web_server->tcp_clients[client_idx]);
+    tcp_server->write_connection(client_idx, ret_data.buff, ret_data.size);
+  }else{ // if insertion failed, it's not in the cache, so just send the original buffer
+    tcp_server->write_connection(client_idx, std::move(buff)); // this works because the rvalue reference of buff isn't assigned to anywhere in try_insert_item (since it failed), so buff still has its data
+  }
 }
 
 template<server_type T>
