+++ test_code.js
@@ -8,5 +8,5 @@ function test_websocket_wait_recv(){
 function test_websocket_send(){
   x = new WebSocket("wss://radio.erewhon.xyz/ws");
   x.onmessage = (e) => e.data.text().then((data) => console.log(data))
-  x.send("Test string");
+  x.onopen = () => x.send("Test string");
 }
